moved webserver init/teardown to a provider
need to setup the request/repsonse logic whenever the server is started or stopped.
